de/exam/requirements.txt
apache-airflow==3.0.2
numpy==2.3.0
pandas==2.3.0
python-dotenv==1.1.0
scikit_learn==1.7.0
